refactor: Minor structure modification
