+++ src/api.ts
@@ -1,8 +1,8 @@
 export { ThemeManager, Theme } from './theme/theme';
 export { Strategy } from './theme/strategy';
 export { Variable } from './theme/variable';
-export { ColorScheme } from './theme/colors/color';
-export { Color, VariableColor } from './theme/colors/color';
+export { ColorScheme } from './theme/color';
+export { Color, VariableColor } from './theme/color';
 export {
   ColorVariant,
   OpacityVariant,
@@ -12,4 +12,4 @@ export {
   VariantTransformer,
   VariantsObject,
   MappedVariant,
-} from './theme/colors/color';
+} from './theme/color';
