+++ src/util/isColorInput.ts
